requirements.txt
Flask
pandas
openpyxl
Flask-Cors
requests
Office365-REST-Python-Client